ci/requirements.txt
hera-workflows==4.0.1
dvc[gs]==2.30.0
pyaml==21.10.1